Add DenseVectorSourceValueFetcher
